# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ReplayGate Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /runs:
    get:
      summary: List runs
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
        - in: query
          name: verdict
          schema:
            type: string
        - in: query
          name: runner
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: since
          schema:
            type: string
        - in: query
          name: until
          schema:
            type: string
      responses:
        '200':
          description: A list of runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
    post:
      summary: Create a run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateRequest'
      responses:
        '201':
          description: Run created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
  /runs/metrics:
    get:
      summary: List run metrics
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Metrics list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunMetricsResponse'
  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        '200':
          description: Prometheus metrics output
          content:
            text/plain:
              schema:
                type: string
  /settings:
    get:
      summary: Get settings
      responses:
        '200':
          description: Settings response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
    put:
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdateRequest'
      responses:
        '200':
          description: Settings response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
  /runs/{id}:
    get:
      summary: Get run detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          description: Run not found
  /runs/{id}/verdict:
    get:
      summary: Get run verdict
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run verdict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverallVerdict'
        '404':
          description: Verdict not found
  /runs/{id}/artifacts:
    get:
      summary: List run artifacts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactListResponse'
  /runs/{id}/artifacts/{name}:
    get:
      summary: Download artifact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact file
  /runs/{id}/cleanup:
    post:
      summary: Cleanup run side effects
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cleanup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanupResponse'
components:
  schemas:
    RunListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    RunCreateRequest:
      type: object
      required:
        - name
        - recording_id
        - baseline_base_url
        - candidate_base_url
      properties:
        name:
          type: string
        recording_id:
          type: string
        baseline_base_url:
          type: string
        candidate_base_url:
          type: string
        baseline_version:
          type: string
          default: v1
        candidate_version:
          type: string
          default: v2
        baseline_run_id:
          type: string
          nullable: true
        executor:
          type: string
        runners:
          type: array
          items:
            type: string
          default: [replay]
        executor:
          type: string
          default: local
        strict_tolerance:
          type: number
          default: 0.05
        rules:
          $ref: '#/components/schemas/DiffRules'
        thresholds:
          $ref: '#/components/schemas/GateThresholds'
    Run:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        recording_id:
          type: string
        baseline_base_url:
          type: string
        candidate_base_url:
          type: string
        baseline_version:
          type: string
        candidate_version:
          type: string
        baseline_run_id:
          type: string
          nullable: true
        runners:
          type: array
          items:
            type: string
        runner_results:
          type: array
          items:
            $ref: '#/components/schemas/RunnerResult'
        overall_verdict:
          type: string
          enum: [PASS, FAIL]
        strict_tolerance:
          type: number
        rules:
          $ref: '#/components/schemas/DiffRules'
        thresholds:
          $ref: '#/components/schemas/GateThresholds'
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        diff_summary:
          $ref: '#/components/schemas/DiffSummary'
        verdict:
          $ref: '#/components/schemas/Verdict'
        error_message:
          type: string
          nullable: true
    DiffRules:
      type: object
      properties:
        global_ignore:
          type: array
          items:
            type: string
        endpoint_rules:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EndpointRule'
        numeric_tolerance:
          type: number
    EndpointRule:
      type: object
      properties:
        ignore:
          type: array
          items:
            type: string
        strict:
          type: array
          items:
            type: string
        numeric_tolerance:
          type: number
    DiffSummary:
      type: object
      properties:
        diff_rate:
          type: number
        total_fields:
          type: integer
        diff_fields:
          type: integer
        strict_mismatches:
          type: integer
        strict_max_drift:
          type: number
        strict_tolerance_used:
          type: number
        schema_breaking:
          type: integer
        numeric_drift:
          type: integer
        total_requests:
          type: integer
    GateThresholds:
      type: object
      properties:
        replay:
          $ref: '#/components/schemas/ReplayThresholds'
        perf:
          $ref: '#/components/schemas/PerfThresholds'
        security:
          $ref: '#/components/schemas/SecurityThresholds'
        compat:
          $ref: '#/components/schemas/CompatThresholds'
        obs:
          $ref: '#/components/schemas/ObsThresholds'
        max_diff_rate:
          type: number
          default: 0.05
        max_schema_breaking:
          type: integer
          default: 0
        max_strict_mismatches:
          type: integer
          default: 0
    ReplayThresholds:
      type: object
      properties:
        max_diff_rate:
          type: number
          default: 0.05
        max_schema_breaking:
          type: integer
          default: 0
        max_strict_mismatches:
          type: integer
          default: 0
    PerfThresholds:
      type: object
      properties:
        max_error_rate_pct:
          type: number
          default: 0.5
        max_p99_ms:
          type: number
          default: 500
        vus:
          type: integer
          default: 5
        duration:
          type: string
          default: 5s
    ObsThresholds:
      type: object
      properties:
        max_error_rate_pct:
          type: number
          default: 0.5
        max_p99_ms:
          type: number
          default: 500
        window:
          type: string
          default: run
    SecurityThresholds:
      type: object
      properties:
        max_high:
          type: integer
          default: 0
        max_medium:
          type: integer
          default: 0
    CompatThresholds:
      type: object
      properties:
        max_breaking_changes:
          type: integer
          default: 0
        mode:
          type: string
          default: strict
    Verdict:
      type: object
      properties:
        verdict:
          type: string
          enum: [PASS, FAIL]
        reasons:
          type: array
          items:
            $ref: '#/components/schemas/VerdictReason'
    RunnerResult:
      type: object
      properties:
        name:
          type: string
        verdict:
          type: string
          enum: [PASS, FAIL]
        reasons:
          type: array
          items:
            $ref: '#/components/schemas/VerdictReason'
        artifacts_files:
          type: array
          items:
            type: string
        metrics:
          type: object
    OverallVerdict:
      type: object
      properties:
        overall_verdict:
          type: string
          enum: [PASS, FAIL]
        runner_results:
          type: array
          items:
            $ref: '#/components/schemas/RunnerResult'
    VerdictReason:
      type: object
      properties:
        domain:
          type: string
        rule_or_metric:
          type: string
        observed:
          type: string
        threshold:
          type: string
        time_window:
          type: string
        evidence_link:
          type: string
        baseline_run_id:
          type: string
    ArtifactListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactItem'
    ArtifactItem:
      type: object
      properties:
        name:
          type: string
        size_bytes:
          type: integer
        download_url:
          type: string
    RunMetricsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RunMetric'
        summary:
          type: object
          properties:
            pass:
              type: integer
            fail:
              type: integer
    RunMetric:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
        overall_verdict:
          type: string
        p99_ms:
          type: number
          nullable: true
        error_rate_pct:
          type: number
          nullable: true
        rps:
          type: number
          nullable: true
    ThresholdTemplate:
      type: object
      properties:
        name:
          type: string
        thresholds:
          type: object
    SettingsUpdateRequest:
      type: object
      properties:
        default_executor:
          type: string
        threshold_templates:
          type: array
          items:
            $ref: '#/components/schemas/ThresholdTemplate'
        active_template:
          type: string
    SettingsResponse:
      type: object
      properties:
        default_executor:
          type: string
        threshold_templates:
          type: array
          items:
            $ref: '#/components/schemas/ThresholdTemplate'
        active_template:
          type: string
        env:
          type: object
    CleanupResponse:
      type: object
      properties:
        status:
          type: string
        cleaned_at:
          type: string
          format: date-time